Lots of small last minute fixes
git-svn-id: http://agilefant.googlecode.com/svn/trunk@626 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ src/fi/hut/soberit/agilefant/web/UserAction.java
@@ -39,6 +39,11 @@ public class UserAction extends ActionSupport implements CRUDAction{
 			super.addActionError(super.getText("user.hasLinkedItems"));
 			return Action.ERROR;
 		}
+		/* Prevent the deletion of administrator*/
+		if(userId == 1){
+			super.addActionError("User cannot be deleted");
+			return Action.ERROR;
+		}
 		userDAO.remove(userId);
 		return Action.SUCCESS;
 	}
